changed `word` size for presumed 64bit sized bus

+++ src/lib.c
@@ -1,3 +1,4 @@
+#include <assert.h>
 #include <ctype.h>
 #include <limits.h>
 #include <stdarg.h>
@@ -69,7 +70,7 @@ void snprintf_s(char *restrict str, size_t size, const char *restrict fmt,
   if ((size_t)n >= size) panic("format string too long");
 }
 
-typedef int word; // "word" used for optimal copy speed
+typedef long word; // "word" used for optimal copy speed
 
 #define wsize sizeof(word)
 #define wmask (wsize - 1)
